Time: 2 ms (100.00%), Space: 56 MB (23.31%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -10,6 +10,6 @@ class Solution {
             sum = sum+nums[i]-nums[i-k];
             max = Math.max(max,sum);
         }
-        return (double)max/(double)k;
+        return (double)max/k;
     }
 }
